Fix property detection, remove nint casts, add typed MsgSend overloads, convert New() to ctor
1. Fix property detection bug in CppAstParser: the pre-pass was adding names
   to the same set used by the main pass, causing propertyNames.Add() to
   always return false. Use separate sets for collision detection vs tracking.

2. Fix setter method filtering: check potentialPropertyNames instead of
   def.Properties (which may not have the property yet due to iteration order).

3. Remove IsMethodParamsSupported filter, allowing all param combinations.

4. Update UnwrapParam to pass actual types (uint, float, Bool8, etc.) instead
   of casting everything to nint.

5. Refactor ObjectiveCRuntime: keep hand-written core in ObjectiveCRuntime.cs,
   generate needed MsgSend overloads dynamically into
   ObjectiveCRuntime.Generated.cs based on actual usage.

6. Convert static New() factory method to parameterless constructor for
   concrete classes.

7. Regenerate all binding files with the updated generator.

+++ Metal.NET/Metal/MTLTextureSwizzleChannels.cs
@@ -50,7 +50,7 @@ public class MTLTextureSwizzleChannels : IDisposable
 
     public static MTLTextureSwizzleChannels Make(MTLTextureSwizzle r, MTLTextureSwizzle g, MTLTextureSwizzle b, MTLTextureSwizzle a)
     {
-        MTLTextureSwizzleChannels result = new(ObjectiveCRuntime.MsgSendPtr(s_class, MTLTextureSwizzleChannelsSelector.MakeGBA, (nint)(uint)r, (nint)(uint)g, (nint)(uint)b, (nint)(uint)a));
+        MTLTextureSwizzleChannels result = new(ObjectiveCRuntime.MsgSendPtr(s_class, MTLTextureSwizzleChannelsSelector.MakeGBA, (uint)r, (uint)g, (uint)b, (uint)a));
 
         return result;
     }
